+++ src/components/App.tsx
@@ -1,12 +1,13 @@
 import { ChangeEvent, useState, FC } from "react";
 import styled from "styled-components";
 import { MemoList } from "./MemoList";
+import { useMemoList } from "../hooks/useMemoList";
 
 export const App: FC = () => {
+  // カスタムフックからそれぞれ取得
+  const { memos, addTodo, deleteTodo } = useMemoList();
   // テキストボックスState
   const [text, setText] = useState<string>("");
-  // メモ一覧State
-  const [memos, setMemos] = useState<string[]>([]);
 
   // テキストボックス入力時に入力内容をStateに設定
   const onChangeText = (e: ChangeEvent<HTMLInputElement>) =>
@@ -14,11 +15,8 @@ export const App: FC = () => {
 
   // [追加]ボタンを押下時
   const onClickAdd = () => {
-    // State変更を正常に検知させるため新しい配列を生成
-    const newMemos = [...memos];
-    // テキストボックスの入力内容をメモ配列に追加
-    newMemos.push(text);
-    setMemos(newMemos);
+    // カスタムフックのメモ追加ロジック実行
+    addTodo(text);
     // テキストバックスを空に
     setText("");
   };
@@ -26,13 +24,10 @@ export const App: FC = () => {
   // [削除]ボタン押下時（何番目が押されたかを引数で受け取る）
   const onClickDelete = useCallback(
     (index: number) => {
-      // State変更を正常に検知させるため新たな配列を生成
-      const newMemos = [...memos];
-      // メモ配列から該当の要素を削除
-      newMemos.splice(index, 1);
-      setMemos(newMemos);
+      // カスタムフックのメモ削除ロジックの実行
+      deleteTodo(index);
     },
-    [memos]
+    [deleteTodo]
   );
 
   return (
